requirements.txt
fastapi==0.89.0
uvicorn==0.21.1
boto3==1.28.2
python-dotenv==1.0.0
python-multipart==0.0.5
